using app option isCapApplication for check

+++ packages/fiori-elements-writer/src/index.ts
@@ -114,8 +114,8 @@ async function generate<T extends {}>(basePath: string, data: FioriElementsApp<T
 
     // Extend ui5-local.yaml
     const ui5LocalConfigPath = join(basePath, 'ui5-local.yaml');
-    if(fs.exists(ui5LocalConfigPath)) {
-        // ui5LocalConfigPath will be present only if app is not a CAP application
+    if(!feApp.appOptions?.isCapApplication && fs.exists(ui5LocalConfigPath)) {
+        // write ui5-local.yaml only if not a CAP application and it exists
         const ui5LocalConfig = await UI5Config.newInstance(fs.read(ui5LocalConfigPath));
         ui5LocalConfig.addUI5Libs([ushellLib]);
         fs.write(ui5LocalConfigPath, ui5LocalConfig.toString());
